refactor: Refactor schema part 1
