+++ src/components/dashboard-resource/DashboardResourceControls.tsx
@@ -2,7 +2,7 @@
 
 import { Dashboard } from '@/schemas/configuration/dashboardsSchema';
 import EmbedModal from './EmbedModal';
-import { Resource } from '@/schemas/configurationSchema';
+import { Resource } from '@/schemas/configuration/configurationSchema';
 import { createSharedPathnamesNavigation } from 'next-intl/navigation';
 import { locales } from '@/locales';
 import { useSearchParams } from 'next/navigation';
