+++ src/components/dashboard-resource/dashboard/DashboardContent.tsx
@@ -2,7 +2,7 @@ import { categoryToDataElement, dashboardToDataElement, resourceToDataElement } 
 
 import { CSSProperties } from 'react';
 import CarouselContent from './CarouselContent';
-import { Configuration } from '@/schemas/configurationSchema';
+import { Configuration } from '@/schemas/configuration/configurationSchema';
 import { DashboardContent as Content } from '@/schemas/configuration/dashboardsSchema';
 import { DataElement } from '@/types/data';
 import ErrorComponent from '../../error-handling/ErrorComponent';
